Atividades da Primeira aula de C#

+++ aula1exercicios/Atividade 5/Program.cs
@@ -1,31 +1,31 @@
-﻿using System;
+﻿  using System;
 
-namespace Aula1 {
-    class Program {
-        static void Main (string[] args) {
+  namespace Aula1 {
+      class Program {
+          static void Main (string[] args) {
 
-            int num1, num2, num3;
-            string Fulano;
-            float resultado;
+              float num1, num2, num3;
+              string Fulano;
+              float resultado;
 
-            Console.WriteLine ("Atividade1");
-            Console.Write ("Digite seu nome: ");
+              Console.WriteLine ("Atividade1");
+              Console.Write ("Digite seu nome: ");
 
-            Fulano = Console.ReadLine ();
+              Fulano = Console.ReadLine ();
 
-            Console.WriteLine (Fulano + ", Agora digite o valor do seu salário fixo: ");
-            num1 = int.Parse (Console.ReadLine ());
+              Console.WriteLine (Fulano + ", Agora digite o valor do seu salário fixo: ");
+              num1 = float.Parse (Console.ReadLine ());
 
-            Console.WriteLine (Fulano + ", Agora digite o valor das suas vendas: ");
-            num2 = int.Parse (Console.ReadLine ());
+              Console.WriteLine (Fulano + ", Agora digite o valor das suas vendas: ");
+              num2 = float.Parse (Console.ReadLine ());
 
-            Console.WriteLine (Fulano + ", Agora digite o valor da quantidade de vendas esse mês: ");
-            num3 = int.Parse (Console.ReadLine ());
+              Console.WriteLine (Fulano + ", Agora digite o valor da quantidade de vendas esse mês: ");
+              num3 = float.Parse (Console.ReadLine ());
 
-            resultado = num1 + (num2 * num3) * 10 / 100;
+              resultado = num1 + (num2 * num3) * 10 / 100;
 
-            Console.WriteLine ("Seu salário atual é: " + resultado);
+              Console.WriteLine ("Seu salário atual é: " + resultado);
 
-        }
-    }
-}
+          }
+      }
+  }
